Rename onboarding2 -> onboarding (#1083)
