+++ src/onboarding/Configure/ConfigureStepsItem.js
