+++ src/onboarding/Configure/ConfigureStepsPanel.js
