+++ src/onboarding/Deployment/DeploymentStepsPanel.js
