+++ src/onboarding/Deployment/assets/illustration-all-done.png
