+++ src/onboarding/Onboarding/OnboardingTopBar.js
