+++ src/onboarding/Templates/TemplateCard.js
