+++ src/onboarding/Templates/TemplateDetails.js
