+++ src/onboarding/Welcome/assets/action-create.png
