+++ src/onboarding/index.js
