+++ src/onboarding/styles.js
