+++ src/onboarding/symbols.js
