+++ src/onboarding/validate-requirements.js
